+++ public/sw.js
@@ -1,4 +1,4 @@
-const OED_VERSION = '2.0.20200406.0';
+const OED_VERSION = '2.0.20200407.0';
 const OED_BASE = 'OED';
 const ACE_VERSION = '1.4.8';
 const ACE_BASE = 'Ace';