+++ public/res/CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## Version 2.0.30200408 (dev)
+- Improved Close file.
+
 ## Version 2.0.30200407 (dev)
 - Added ChangeEditorOptionEvent class.
 - Changed class name from View to AppView.